update map position tracking
Relationship between MAP_X and horizontal scroll is now inverse, so
moving to the right increases MAP_X and decreases horizontal scroll
so the level moves to the left as the "camera" moves to the right.

MAP_X starts at a magic positive value which evaluates to the
same horizontal scroll value for Plane A that is in VRAM at the end
of the opening cinematic.

MAP_X is still fixed point, and is shifted right 4 times and negated
before being applied to horizontal scroll.

To see if new parts of the level need to be loaded/unloaded, there is
now a MAP_X_OLD and MAP_Y_OLD to check if a player has crossed a
boundary where stuff needs to be loaded.